Moves NeuralNet to domain
