+++ src/indiv.h
@@ -8,7 +8,9 @@
 #include <array>
 #include "common/include/coord.h"
 #include "common/include/dir.h"
-#include "genome-neurons.h"
+// #include "genome-neurons.h"
+#include "sensors-actions.h"
+#include "domain/include/neuralNet.h"
 
 namespace BS {
